Proyecto funcional, se aplica la toda vista en mainwindow

+++ RedSendero/src/model/Estacion.java
@@ -1,25 +1,25 @@
 package model;
 
 public class Estacion {
-    private String nombre;
-    private int x, y;  // Coordenadas para dibujar en el panel
+	private String nombre;
+	private double x, y; // Coordenadas para dibujar en el panel
 
-    public Estacion(String nombre, int x, int y) {
-        this.nombre = nombre;
-        this.x = x;
-        this.y = y;
-    }
+	public Estacion(String nombre, double x, double y) {
+		this.nombre = nombre;
+		this.x = x;
+		this.y = y;
+	}
 
-    // Getters
-    public String getNombre() { 
-    	return nombre; 
-    	}
-    
-    public int getX() {
-    	return x;
-    }
-    
-    public int getY() { 
-    	return y;
-    }
+	// Getters
+	public String getNombre() {
+		return nombre;
+	}
+
+	public double getX() {
+		return x;
+	}
+
+	public double getY() {
+		return y;
+	}
 }
